added automted bundle single program file
The first quick script testing pyaudacity and mutagen to mix tracks. Born in a few hours of frustration with the tedium of manual mixing and now the target of a revamp with better design deepening its core features, ease of use, and leveraging more oop principles and error handling to mitigate bugginess and unreliability during program execution.